优化系统升级

+++ src/main/java/com/benzolamps/dict/service/impl/VersionServiceImpl.java
@@ -5,20 +5,18 @@ import com.benzolamps.dict.service.base.VersionService;
 import com.benzolamps.dict.util.Constant;
 import com.benzolamps.dict.util.DictFile;
 import lombok.extern.slf4j.Slf4j;
+import lombok.var;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.UrlResource;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
-import org.springframework.util.FileCopyUtils;
 import org.springframework.util.FileSystemUtils;
 import org.springframework.util.StreamUtils;
 
+import javax.annotation.PostConstruct;
 import javax.annotation.Resource;
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.InputStream;
-import java.io.OutputStream;
+import java.io.*;
 import java.util.List;
-import java.util.Map;
 import java.util.function.Consumer;
 
 /**
@@ -32,16 +30,28 @@ import java.util.function.Consumer;
 @Slf4j
 public class VersionServiceImpl implements VersionService {
 
-    private long size;
+    @Value("${install.total:}")
+    private int total;
 
+    @Value("${install.totalSize:}")
+    private long totalSize;
+
+    @Value("${install.lastDelta:}")
     private long lastDelta;
 
-    private Map versionInfo;
+    @Value("${install.succeed:}")
+    private Boolean succeed;
+
+    private boolean dead;
+
+    private List<String> versionInfo = Constant.EMPTY_LIST;
+
+    private String newVersionName;
 
     @Value("#{dictProperties.remoteBaseUrl}")
     private String baseUrl;
 
-    private Status status;
+    private Status status = Status.ALREADY_NEW;
 
     @Resource
     private DictProperties dictProperties;
@@ -55,58 +65,78 @@ public class VersionServiceImpl implements VersionService {
 
     @Override
     public void update() {
-        size = 0;
-        lastDelta = 0;
-        if (thread == null) {
-            thread = new Thread(() -> {
-                try {
-                    updateProcess();
-                } catch (Throwable e) {
-                    logger.error(e.getMessage(), e);
-                    status = Status.FAILED;
-                    callback.accept(status);
-                } finally {
-                    thread = null;
-                }
-            });
-            thread.start();
+        /* 下载完成或者更新完成的状态时不能再次更新 */
+        if (status != Status.DOWNLOADED && status != Status.INSTALLED) {
+
+            /* 开启一个更新线程 */
+            if (thread == null) {
+                (thread = new Thread(() -> {
+                    try {
+                        updateProcess();
+                    } catch (Throwable e) {
+                        logger.error(e.getMessage(), e);
+                        status = Status.FAILED;
+                        callback.accept(status);
+                    } finally {
+                        thread = null;
+                    }
+                })).start();
+            }
         }
     }
 
-    private Map getVersionInfo() {
+    /* 重置版本信息 */
+    @PostConstruct
+    @Scheduled(fixedRate = 1000 * 15)
+    private void resetVersionInfo() {
+
+        if (succeed != null && status == Status.ALREADY_NEW) {
+            status = succeed ? Status.INSTALLED : Status.FAILED;
+            return;
+        }
+
         try {
-            UrlResource resource = new UrlResource(baseUrl + "/yml/version.yml");
-            return Constant.YAML.loadAs(resource.getInputStream(), Map.class);
+            /* 下载完成或者更新完成的状态时不能再次更新 */
+            if (status != Status.DOWNLOADED && status != Status.INSTALLED) {
+                UrlResource resource = new UrlResource(baseUrl + "/yml/version.yml");
+                versionInfo = Constant.YAML.loadAs(resource.getInputStream(), List.class);
+
+                /* 判断是否有新版本 */
+                if (versionInfo.stream().anyMatch(str ->
+                    (newVersionName = getVersion(str)) != null &&
+                    newVersionName.compareTo(dictProperties.getVersion()) > 0)) {
+                    status = Status.HAS_NEW;
+                } else {
+                    newVersionName = dictProperties.getVersion();
+                }
+            }
         } catch (Throwable e) {
+            if (newVersionName == null) newVersionName = dictProperties.getVersion();
             logger.error(e.getMessage(), e);
-            return null;
         }
     }
 
     @Override
     public Status getStatus() {
-        if (status == Status.ALREADY_NEW || status == Status.HAS_NEW || status == Status.FAILED || status == null) {
-            versionInfo = getVersionInfo();
-            if (versionInfo == null || dictProperties.getVersion().equals(versionInfo.get("name"))) {
-                return status = Status.ALREADY_NEW;
-            } else {
-                return status = Status.HAS_NEW;
-            }
+       return status;
+    }
+
+    /* 从文件名中提取版本号 */
+    private String getVersion(String str) {
+        if (str == null || !(str = str.toLowerCase()).contains("-") || !str.endsWith(".zip")) {
+            return null;
         }
-        return status;
+        return str.substring(str.lastIndexOf("-") + 1, str.indexOf(".zip"));
     }
 
     @Override
     public String getTotalSize() {
-        return DictFile.fileSizeStr(size);
+        return DictFile.fileSizeStr(totalSize);
     }
 
     @Override
     public int getTotal() {
-        if (versionInfo != null) {
-            return ((List)versionInfo.get("files")).size();
-        }
-        return 0;
+        return total;
     }
 
     @Override
@@ -116,49 +146,63 @@ public class VersionServiceImpl implements VersionService {
 
     @Override
     public String getNewVersionName() {
-        if (versionInfo != null) {
-            return (String) versionInfo.get("name");
-        }
-        return null;
+        return newVersionName;
     }
 
     private void updateProcess() throws Throwable {
-        long start = System.currentTimeMillis();
-        versionInfo = getVersionInfo();
+
         if (versionInfo != null) {
-            List<String> fileNames = (List<String>) versionInfo.get("files");
-            File tempFile = new File(tempPath);
-            tempFile.mkdirs();
+
+            /* 重置各种计数器 */
+            totalSize = 0; lastDelta = 0; total = 0; dead = false;
+
+            long start = System.currentTimeMillis();
+
+            /* 更改状态 */
             status = Status.DOWNLOADING;
             callback.accept(status);
-            for (String fileName1 : fileNames) {
-                String url = baseUrl + "/" + fileName1;
-                UrlResource resource = new UrlResource(url);
-                InputStream inputStream = resource.getInputStream();
-                File file = new File(tempPath + "/" + fileName1);
-                if (file.getParentFile() != null) file.getParentFile().mkdirs();
-                OutputStream outputStream = new FileOutputStream(file);
-                StreamUtils.copy(inputStream, outputStream);
-                inputStream.close();
-                outputStream.close();
+
+            /* 创建临时路径 */
+            File tempFile = new File(tempPath);
+            if (tempFile.exists()) {
+                if (tempFile.isDirectory()) FileSystemUtils.deleteRecursively(tempFile);
+                else tempFile.delete();
             }
-            status = Status.COPYING;
-            callback.accept(status);
-            for (String fileName : fileNames) {
-                File from = new File(tempPath + "/" + fileName);
-                File to = new File(fileName);
-                if (to.getParentFile() != null) to.getParentFile().mkdirs();
-                to.createNewFile();
-                FileCopyUtils.copy(from, to);
-                size += to.length();
+            tempFile.mkdirs();
+
+            /* 下载文件 */
+            for (String fileName : versionInfo) {
+
+                String version = getVersion(fileName);
+
+                /* 判断是不是新版本 */
+                if (version != null && version.compareTo(dictProperties.getVersion()) > 0) {
+
+                    /* 生成本地路径 */
+                    String url = baseUrl + "/" + fileName;
+                    UrlResource resource = new UrlResource(url);
+                    File file = new File(tempPath + "/" + fileName);
+
+                    /* 创建父路径 */
+                    if (file.getParentFile() != null && !file.getParentFile().exists()) file.getParentFile().mkdirs();
+
+                    /* 创建新文件 */
+                    file.createNewFile();
+
+                    /* 流复制 */
+                    try (var is = resource.getInputStream(); var os = new FileOutputStream(file)) {
+                        int size = StreamUtils.copy(is, os);
+                        this.totalSize += size;
+                        ++total;
+                    }
+                }
             }
-            status = Status.DELETING;
-            callback.accept(status);
-            FileSystemUtils.deleteRecursively(tempFile);
+
             long end = System.currentTimeMillis();
             lastDelta = end - start;
-            status = Status.FINISHED;
-            dictProperties.setVersion(getNewVersionName());
+
+            /* 更改状态 */
+            status = Status.DOWNLOADED;
             callback.accept(status);
         }
     }
@@ -169,7 +213,12 @@ public class VersionServiceImpl implements VersionService {
     }
 
     @Override
-    public void resetStatus() {
-        status = Status.ALREADY_NEW;
+    public void die() {
+        dead = true;
+    }
+
+    @Override
+    public boolean isDead() {
+        return dead;
     }
 }
